Allow large numbers of draw objects (#526)
* Allow large numbers of draw objects

Previously there was a limit of workgroup size squared for the number of draw objects, which is 64k in practice. This PR makes each workgroup iterate multiple blocks if that limit is exceeded, borrowing a technique from FidelityFX sort.

WIP, this causes hangs on mac. Uploading to test on other hardware.

Also contains some changes for testing that may not want to be committed as is.

Fixes #334

* Add missing barrier

Add barrier for write-after-read hazard in coarse. The loop in question processes 64k draw objects at a time, so the barrier only gets invoked when that limit is exceeded.

Also move new test scene so it isn't the first.

* Address review comments

Set resolution in params for test scene. Add comments explaining division of work.