Add robustness to GPU shaders (#537)
* Add robustness to GPU shaders

Make each stage quit early if a previous stage has failed.

The CPU shaders are minimally changed to be layout compatible. For the most part, they'll panic on a bounds check if sizes are exceeded. That's arguably useful for debugging, but a case can be made they should have the same behavior as the GPU shaders.

Work towards #366

* Address review feedback

Clarify some nits, and also make a distinction between reporting failure in path_count and coarse.