[fine] Define the "r8" permutation
The "r8" variant of the fine stage can be used to target storage
textures with the r8unorm texel format. This is specifically intended
to reduce the texture memory required for coverage-mask renderers that
only output 8-bits of coverage-as-alpha in a single channel. This
format is non-standard and requires the an implementation-specific
extension.

Currently this usage requires that any encoded color in a path brush use
only the red channel as the remaning channels will be ignored.