[fine] Coverage-mask specialization and some shader feature cleanup (#540)
* [shaders] Remove force_rw_storage feature

This feature is no longer used and can be removed

* [shaders] Repurpose the "full" feature

Previously the "full" feature was introduced to distinguish the
complete imaging model of vello from the more limited "reduced"
variant, which was intended to produce coverage masks for an atlas
based path renderer. This reduced variant was limited to only paths
without any blends, clips, or other non-path draw objects. This
specialization wasn't fully carried forward with the move from piet-gpu
to Vello, since some of the excluded features (in particular
anti-aliased clipping) turned out to be a desirable feature for
atlasing.

While there is still room for pipeline specializations and optimizations
geared towards coverage-mask atlas based path renderers, the existing
meaning of "full" left some important features out (such as the PTCL and
stroke style / path_ix monoids used in stroke expansion).

This change re-purposes the meaning of the "full" specialization. It is
now only present in the fine rasterization stage and enables the
gradient commands.

The vello_shaders crate now also provides a (default) feature called
"full" that gates the "full" define in all shader stages.

* [fine] Define the "r8" permutation

The "r8" variant of the fine stage can be used to target storage
textures with the r8unorm texel format. This is specifically intended
to reduce the texture memory required for coverage-mask renderers that
only output 8-bits of coverage-as-alpha in a single channel. This
format is non-standard and requires the an implementation-specific
extension.

Currently this usage requires that any encoded color in a path brush use
only the red channel as the remaning channels will be ignored.