[backdrop_dyn] Handle upstream pipeline failure
Following #537 it is possible for the flatten stage to fail and flag
a failure. In some cases this can cause invalid / corrupt bounding box
data to propagate downstream, leading to a hang in the per-tile
backdrop calculation loop.

Triggering this is highly subtle, so I don't have a test case as part of
vello scenes that can reliably reproduce this. Regardless, it makes
sense to check for the upstream failures and terminate the work in
general.