Use vello_shaders crate to load and preprocess WGSL (#563)
This replaces the shader preprocessing code in the vello crate with
invocations to the vello_shaders crate, which supports both reading from
a directory at runtime (for hot_reload) and bundlng pre-compiled shader
code into the binary (which is suitable for wasm).

This also moves the CPU shader code over to the vello_shaders crate
behind the "cpu" feature.

Resolves #467