Remove download command, keep download directory.
This still allows loading from the download directory, but removes
the command that would download to it, allowing people to populate
it themselves.

The files downloaded were parsed with the `vello_svg` parser, but
that was removed from here and moved to its own home, so files
loaded would need some preprocessing (to be defined elsewhere).