Better error types (#516)
Using custom `thiserror` errors instead of `Box<dyn Error>`, closes #285

This removed the use of `Result` as a return type for many functions,
including `RenderContext::new()`. This "could" fail when no backend
feature is activated for wgpu, but because we don't vendor them, this
should never happen. (also there is no failable version of this
function)

I also had to make a `RendererError` which I totally hate, but the
problem here is that without the `wgpu-profiler` feature this function
never fails, so should I remove the `Result` and just unwrap for the
`GpuProfiler`?

There are still many unwraps, which may(?) still be handled correctly
with errors, but for most of them I'm not sure. They are in many
functions which had `Result` return types but still, `unwrap` was
called, so was it intention?

---------

Co-authored-by: Kaur Kuut <strom@nevermore.ee>