Workaround to avoid full hang (#589)
As the scale factor becomes unreasonably large, the number of
subdivisions of an Euler spiral into lines grows, making the flatten
stage take too long.

This patch just bounds the number of subdivisions, so it will eventually
make progress. It will be slow, but not hang.

A better solution would be to aggressively cull so it only generates
geometry inside the viewport, but that is considerably more complicated.

Workaround for #548