Migrate the repository to a virtual workspace. (#590)
The CI, VSCode config, `ARCHITECTURE.md` and some `README.md` links have
been updated to reflect the new directory structure.

The WGSL shaders now live in the `vello_shaders` subdirectory. Vello hot
reloading depends on our repo structure, which is fine as decided at
last office hours due to this being a development feature.

The examples continue to live in a root `examples` directory, as that is
how it remained for [Parley](https://github.com/linebender/parley) as
well. I did this for consistency and don't really feel strongly either
way if people think we should move them under `vello/examples` instead.
If we had examples without `Cargo.toml` that would make more sense, but
we don't, so ..

The repo `README.md` and the `vello` sub-directory `README.md` are exact
clones and both need updating. That is intentionally planned as
follow-up work to reduce the review friction here.