Add inner joins
Draw inner join for almost all joins. The exception is when the join
angle is tiny. It shouldn't be too hard to add optimizations, especially
when it's possible to evaluate that the join is entirely inside one side of the stroke.

It's under an #ifdef, the old code is not changed substantially.