[pico_svg] Encode group transform in scene
Build the scene recursively, applying the transform at group nodes, rather than flattening to a single list of items. This causes the transform to be applied correctly to stroke styles, which among other things fixes the rendering of the waves example from the Nehab timings data set.

We're deliberately keeping things minimal, but if we were to add other properties to groups (clips, opacity, etc), then this change would support that.