Add some snapshot test infrastructure (#610)
Fixes #608 

This does not add a wide gamut of tests - I imagine that the
`test_scenes` will provide some good fodder here.
There are some quality-of-life features included:

- You can use `VELLO_TEST_CREATE=all` to create new tests. It is
recommended to set this on your local machine (and it won't be set on
CI)
- You can use `VELLO_TEST_UPDATE=all` to update with new contents if the
test fails. (Note that if there's a minor change which passes the test,
this won't update)

These variables can also filter to specific tests if needed.
The saved snapshots are always created using the GPU renderer, but it is
set up to be easy to use the CPU renderer for tests as well.

Using `cargo nextest` for running tests is recommended; indeed, it seems
to be now required, as the tests segfault without it