Prepare the changelog for 0.2.1 (#632)
This release is for:
- #630

This will also include:
- #631
- #619

Note that this will be a cherry-picked patch release from 0.2.0, so
doesn't have all the changes there.
This PR is to update the changelog on main, and will also be
cherry-picked into the release branch. I will make a release branch and
get sign-off on that before publishing.