[Experiment] Store snapshots in `git lfs` (#643)
This a living experiment to determine how much LFS bandwidth we would
use if we stored our snapshot tests in Git LFS.

The intention is once this is merged, we will then add some additional
snapshot tests, which use this new folder, but which won't fail (at
least not on CI?) if LFS downloading failed. That way, we have a
graceful degradation if we do run out of bandwidth, but improving
resilience if LFS is working.