Allow deeper blend stacks by spilling to a buffer (#657)
This brings in support for blend spilling (which was supported in the
old piet-gpu).

I don't have a good heuristic for how big to make the buffer. That is
something which will need to be addressed in #606 (or its successors). I
just guessed that 256 spills would be fine. I think this is probably too
small - I suspect we'll get feedback from @TrueDoctor about this.

I have confirmed that the robustness works as expected with the GPU
shaders.