Add test scene for clipping
Add many_clips test scene and snapshot test.

Note that doing more than 128 clips, while it would exceed the 256 clip object limit (counting both begin and end clip), would not trigger the failure, as the end clip would be on a workgroup boundary. Using multiples of 3 (as in the original report) triggers the bug, and a full system hang was verified with the fix disabled.