Add SDL2 & C-powered window handling compatibility. (#671)
This pull request adds `create_surface_unsafe()`, that allows you to add
unsafe surface targets.

This allows Vello to be used in SDL2, and it should also work in other
libraries like GLFW, but that's not tested yet.

Here's a screenshot of Vello running in SDL2, i'll push an example of it
soon :D.


![image](https://github.com/user-attachments/assets/f968ea22-bbcd-4408-91d4-2e8c15b2df20)

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>