Make tangents at subpath end watertight (#695)
At subpath end, the last path segment is encoded with the end point
chosen so that the vector from the current point to the end point is the
tangent, ie the same vector as cubic_start_tangent applied to the first
segment in the subpath. In those cases, compute the tangent by
subtracting those two points, rather than cubic_start_tangent applied to
the line. These are mathematically identical, but may give different
results because of roundoff.

There are two cases: an open subpath, in which case the case is applied
to the tangent at draw_start_cap, or a closed subpath, where the logic
is in the tangent calculation in read_neighboring_segment.

Both GPU and CPU shaders are updated. It hasn't been carefully
validated.

Hopefully fixes #616 and #650.