Add `draw_blurred_rounded_rect_in` (intended for box shadows) (#700)
This avoids needing a clip layer for a rounded rectangle which doesn't
need to draw the inner parts (e.g. a transparent background with a box
shadow).

The previous test image was oversized, so this also fixes that (and
makes any future such image fail the tests as expected). I imagine that
the author of that test manually copied the value into the place, not
realising the size of the image.

---------

Co-authored-by: Raph Levien <raph.levien@gmail.com>