Hide the async pipeline from the public API (#706)
In doing ports to Vello 0.3 in our other repositories, I kept running
into that adding the `debug_layers` field needed to be added, and I
don't think it will be clear to users what it's for.

In addition, there is no reason to use the async pipeline for end users
(at least as things stand). In #694, the debug layers feature was marked
as internal, so that isn't a good reason.