Fix inference conflict with `serde_json` through `schemars` feature (#733)
Initially marking as a draft, because this currently just fiddles with
Cargo.toml to ensure that CI tests the feature.
this shouldn't actually require any changes to our CI scripts to
trigger, and as Daniel suggested, this uses a feature on with_winit
rather than vello itself.

Will follow up with the actual fix once I've confirmed CI fails.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>