pico_svg: Simplify color parsing. (#753)
The color parsing in `color` is more complete than the older code in
`peniko`, so this can be simplified now.

Also, switch from `peniko::Color` to `AlphaColor<Srgb>` to continue down
the path of using `peniko::Color` less.

Finally, add a couple of tests to show the color parsing working with a
couple of the supported syntaxes.