Encode `path_data` and `draw_data` directly as `u32` (#817)
We already assume these are word-aligned, and so this would catch a
failure of that assumption very early, if it came up.

A potential alternative is to store `path_data` as f32 instead, as all
values stored are f32. I've not done that for pretty weak reasons, and
it should be trivial to change.

A future change would be to encode the full scene encoding as `u32`.
That would be useful for #810, as the current code has an alignment
hazard.