Add the code for `vello_common` (#828)
Note that this is not ready for merging yet as I'm still tracking down a
regression I introduced. I didn't notice this because in my branch all
affected tests use the rectangle fast path, which I removed here for
now. So that might take a while.

But other than that, this part of the code should be pretty much done
(and reviewable).