Simplify strip generation and analytical coverage calculation
For reference, the code currently in main generates strips for a 200x200
Tiger in 0.475 ms on my hardware (1000 iterations). The code in this PR
generates the strips in 0.359 ms.

This also appears to improve robustness.

The calculations here are ready for lines being removed from tiles,
which is a likely future optimization. Currently, with lines being
packed into tiles, some calculations here do a bit more work than
necessary.

There are two follow-ups to this I'm planning. One is culling tiles to
the left of the viewport, which will allow storing tile x-coordinates
unsigned. The other is to elide horizontal geometry at the
tile-generation stage, as horizontal geometry doesn't impact winding.
That would include generalizing sparse fills to sparse alpha masks,
where the alpha values are invariant over the columns.

As drive-by changes I've improved some comments, and changed some
constants to be associated constants on `Tile` and `WideTile` for easier
importing. As we're still a bit messy with `u16`, `i32`, `usize`, etc.,
some more clean-up will be required in the future changes.