Simplify strip generation and analytical coverage calculation (#835)
The footprint calculation is removed from strip generation. Analytical
pixel coverage and the strip boundaries are now calculated in a single
forward scan through the tiles. This simplification potentially makes it
easier to base SIMD versions on the math here. (As an aside, AVX512
could do a 4x4 tile in `f32` in one go.)

This is a performance improvement: the code in main generates strips for
a 200x200 Tiger in 0.475 ms on my hardware (1000 iterations). The code
in this PR generates the strips in 0.359 ms.

This also appears to improve robustness: for example, the first tiger is
rendered by main, the right tiger is rendered by the code in this PR.


![test-orig](https://github.com/user-attachments/assets/a2a55ed6-ddda-4013-9357-f90e283ab3ed)

![test](https://github.com/user-attachments/assets/ddd142da-a779-4f13-b6d2-78e21b3d6616)

The calculations here are ready for lines being removed from tiles,
which is a likely future optimization. With lines currently being packed
into tiles, some calculations here do a bit more work than necessary.

There are two follow-ups to this I'm planning. One is culling tiles to
the left of the viewport, which will allow storing tile x-coordinates
unsigned. The other is to elide horizontal geometry at the
tile-generation stage, as horizontal geometry doesn't impact winding.
That would include generalizing sparse fills to sparse alpha masks,
where the alpha values are invariant over the columns.

As drive-by changes I've improved some comments, and changed some
constants to be associated constants on `Tile` and `WideTile` for easier
importing. As we're still a bit messy with `u16`, `i32`, `usize`, etc.,
some more clean-up will be required in the future changes.

Some snapshots have some slight rounding differences of colors on some
pixels. The color values of `filled_vertical_hairline_rect.png` match
those calculated by hand, whereas the old snapshot on main does not.