Vello Hybrid: Render into a user-provided RenderPass (#854)
Some of the follow-up from #831.

This strategy might not work as well if/when we start doing clipping and
blending. Things get messy there; ultimately I *suspect* it's still
right to do the "final" render into a user-provided pass, but we want to
do intermediate renders ourselves?
Nevertheless, for the current state this is the correct API; we
definitely should not be doing our own queue submission, anyway.

I've run out of time for today, so have only run the `svg` example
(which worked as expected). I expect the others to work, but I'd
recommend checking that

---------

Co-authored-by: Alex Gemberg <gemberg@canva.com>