Panic instead of silently failing in `vello_hybrid` (#864)
Small changes in `vello_hybrid` to ensure that panics are called when
invariants are not obeyed rather than silently failing.

Implementation like the below means that incorrect use of the API
results in no feedback to the user. We should panic in such cases
because the user has not followed our architectural invariant of
`prepare` being called before `render`.

```rs
fn render (...) {
        let Some(resources) = &self.resources else {
            return;
        };
       ....
```

---

I think we could explore returning some struct from `prepare` that can
be passed to `render` to ensure that prepare must be called before
render, but I leave that as a separate exercise.

---

Tip: Review with whitespace hidden:


![image](https://github.com/user-attachments/assets/1110bab2-c0b8-4516-ad9b-416c42d53390)