`vello_api`: Make `no_std` and add CI (#862)
This makes the workspace dependencies for `kurbo` and `peniko` not
enable default features, and therefore have to enable them in the crates
that use them.

Add CI based on other Linebender crates, except that there is now a
variable to specifically control which targets are checked for `no_std`
compatibility.

This also means that `no_std` crates are checked for MSRV compliance
which was not currently happening for the sparse strips crates.

Other crates will get `no_std` support in the future.