`vello_hybrid`: Fix pixel coord -> NDC transform
I believe the coordinate systems are aligned at the pixel edge, not the
pixel center. This fixes rendering artifacts on my system (AMD GPU),
where with the previous shader code the locations sampled during
fragment shading probably sometimes got bumped to the next pixels.

This should be confirmed to work on systems where it worked previously
as well.