[Sparse Strips]: Take scale from glyph transform and apply it to font size when hinting (#889)
In https://github.com/linebender/vello/pull/883, we discussed
(https://github.com/linebender/vello/pull/883#discussion_r2020456497)
how to handle hinting when the glyph is later transformed. The approach
Vello uses (and shown in this PR): if hinting is requested and the
transform contains a uniform scale (without skew / rotation), extract
the scale from the transform and apply it to the font size and glyph
positions.

This behaviour is the same used by Vello:

1. Extracting and applying scale to font size when hinting is enabled:


https://github.com/linebender/vello/blob/61f82e4e66a9111d6b2ee17f173167c3601c2e55/vello_encoding/src/resolve.rs#L414-L428

2. Applying that scale to glyph positions:


https://github.com/linebender/vello/blob/61f82e4e66a9111d6b2ee17f173167c3601c2e55/vello_encoding/src/resolve.rs#L334

If a consumer doesn't want this behaviour, then they can disable hinting
on scaling.

You can see what effect this has on our snapshot:


https://github.com/user-attachments/assets/9e8577d2-94e8-4722-8878-3a55dc64b6a6