Generalize glyph hinting: allow glyph transforms and global skews (#897)
On top of https://github.com/linebender/vello/pull/896 (that should land
first). See c0b778693cc19d4c3520c9313bd46d7d2641ce7f and onwards for the
changes here.

This builds on the work landed in
https://github.com/linebender/vello/pull/889.

This generalizes our detection of when (vertical-only) hinting can be
performed. When the second coefficient (`b`) of the affine
transformation matrix is 0 (first column, second row), glyphs are not
skewed vertically or rotated, and can be hinted. Because of this
generalization, the `horizontal_skew` special case has been removed.

We now calculate the total transformation matrix once, and separately
calculate the glyph translations using the run's transformation matrix
for global glyph positioning. This reduces the number of affine
multiplies per glyph to one-and-a-bit down from three.