fix(vello_hybrid): update config on resize (#895)
`velly_hybrid` resizing issue was caused by not updating `Config` when
the viewport size changed. As a result, the vertex positions in the
shader were being calculated based on the previous config, which led to
distorted rendered content.

Before:



https://github.com/user-attachments/assets/561e0bb3-408d-4afc-8946-48bc3e98db15



After:


https://github.com/user-attachments/assets/51ac9591-fa67-4e2d-aa66-dff746680547


https://github.com/user-attachments/assets/963497cf-7061-47d8-ba8a-4723e456d4b5


Images in the videos may flicker during resizing because of the problem
seems related to wpu (more details in [#vello > Vello Hybrid: Resizing @
💬](https://xi.zulipchat.com/#narrow/channel/197075-vello/topic/Vello.20Hybrid.3A.20Resizing/near/510114878)):


https://github.com/user-attachments/assets/7feeb008-8f97-42c0-9ab5-da6eb53dec75