Fix issue in filling of clips (#914)
I noticed another issue in filling of clips (which already existed
previously, but unfortunately I didn't detect it in my last PR).
Currently, if I fill a rectangle with a black fill of opacity 0.5 onto a
white background, I get the RGB values 132 134 138 instead of the
expected values 127, 127, 127. I believe the issue was that we
mistakenly multiplied the alpha of the source color to the source, which
is not necessary because the color is already premultiplied.

With this PR, I get the expected values 127, 127, 127. It also fixes an
inconsistency with an existing test, where the color was not consistent
with the stripped part of the rectangle:

<img width="856" alt="image"
src="https://github.com/user-attachments/assets/25d274ad-e66d-4d09-8d70-5b3e24f8c334"
/>


Once gradients are merged, we can probably rewrite `clip_fill` and
`clip_stroke` to reuse the same method as for normal alpha compositing,
to avoid code duplication.